SignalR Web Application Demo Test
I've added new Web Application project. It contains TestPage.html which
shows a map and a red dot simulating the movement of a plane, which is
updated by a SignalR back-end Hub. I am using knockout as a ViewModel
engine to update the position of the dot in the html.